Vanity Generator v1 (and debug mode disabled)
Vanity Generator v1 has been ported from SCC web 3.0 to the DogeCash web wallet, this is an extremely simple VanityGen replica, with the ability to create short custom prefixes up to 3 or 4 letters max currently.

The generator is single-threaded and runs in the renderer process, which means the site will 'lock up' during the generation, this limitation will eventually be bypassed with multi-threaded background workers, but until then, the site will only allow a maximum of 4 custom letters, as higher would currently take a painful amount of time, until further optimized!